+++ overseerr/CHANGELOG.md
@@ -1,5 +1,8 @@
 - Added support for configuring extra environment variables via the `env_vars` add-on option alongside config.yaml. See https://github.com/alexbelgium/hassio-addons/wiki/Add-Environment-variables-to-your-Addon-2 for details.
 
+## 1.34.0-1 (09-01-2026)
+- ⚠ MAJOR CHANGE : switch to the new config logic from homeassistant. Your configuration files will have migrated from /config/addons_config/overseerr to a folder only accessible from my Filebrowser addon called /addon_configs/xxx-overseerr. Please make a backup before updating.
+
 ## 1.34.0 (29-03-2025)
 - Update to latest version from linuxserver/docker-overseerr (changelog : https://github.com/linuxserver/docker-overseerr/releases)